18:18

+++ srcs/ft_putnbr_base.c
@@ -1,4 +1,4 @@
-#include "libft.h"
+#include "../ft_printf.h"
 
 static int	is_valid(char* base, int basenbr)
 {
@@ -22,14 +22,14 @@ static int	is_valid(char* base, int basenbr)
 	return (1);
 }
 
-static void	recur(long int nbr, char* base, int basenbr, size_t* len)
+static void	recur(long long int nbr, char* base, int basenbr, size_t* len)
 {
 	if (nbr >= basenbr)
 		recur(nbr / basenbr, base, basenbr, len);
 	ft_putchar(base[nbr % basenbr], len);
 }
 
-void	ft_putnbr_base(long int nbr, char* base, size_t*len)
+void	ft_putnbr_base(long long int nbr, char* base, size_t*len)
 {
 	int			basenbr;
